Fix: seed logic separation
